adding more functionalities to slash command
